Add page worker feature
Page workers are worker entrypoints. They get bundle split with all
other bundles, and a worker.json file is placed in the output directory
with the page. You can load this file, get the worker filename and
then create a worker from that entrypoint. This feature allows you
to create workers in pages that get bundled with the rest of your site.